Fix intermittent JSONRPC test
* The test was meant to validate that it splits payloads that are too large but it was making an actual HTTP request to validate it.
* This commit adds ByPass to mock the HTTP server and returns exactly the same huge payload at the response.
